1. Socks5 support is fixed in M.S Windows, Mac OS X, Linux and BSD.
[sthp](https://github.com/KaranGauswami/socks-to-http-proxy)'s build
file must be placed in or next to the bundle in Mac and Windows for persepolis bundle. build file must be renamed to
"sthp.exe" in windows and "sthp" for mac.

[Gost](https://github.com/ginuerzh/gost/tree/master) can be used in
Linux and BSD.

[pproxy](https://github.com/qwj/python-proxy) can also be used. PProxy
can be installed by pip.
close #411

2. "<Fault 1: 'No URI to download.'>" Bug is fixed.

This bug is related to aria2c. but it's fixed with a workaround.
Persepolis will restart aria2c, If this situation occures.
close #651
close #723
issue #777

3. Bugs that related to browser extension have been fixed.

New ".py"" file is added to project. Please checkout "resources/PersepolisBI.py".
In M.S.Windows, This file must be frozen(PersepolisBI.exe) by pyinstaller and placed next to the bundle.